Api/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
click==8.1.8
colorama==0.4.6
fastapi==0.115.6
h11==0.14.0
httptools==0.6.4
idna==3.10
loguru==0.7.3
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
watchfiles==1.0.3
websockets==14.1
win32_setctime==1.2.0
